refactor(tools): relocate the Tool ABC to Tools/base.py
The ABC is load-bearing -- risk_tags is what the permission gate resolves
against and every pack tool subclasses it -- while the ToolExecutor beside
it has no callers left and is about to be removed. Separating them means
the deletion cannot take the ABC with it.

tool_executor re-exports Tool so both import paths resolve to the SAME
class object; 9 importers repointed to Tools.base.
